Refactor NextConferencePage to use SpeakerList
