+++ client/styles/main.scss
@@ -37,7 +37,7 @@
 @import '../../shared/components/TicketList/index';
 @import '../../shared/components/OpportunityTickets/index';
 @import '../../shared/components/SpeakersPage/index';
-@import '../../shared/components/SpeakersList/index';
+@import '../../shared/components/SpeakerList/index';
 @import '../../shared/components/ErrorPage404/index';
 @import '../../shared/components/ErrorPage500/index';
 @import '../../shared/components/CommunityLayout/index';
