+++ shared/components/NextConferenceEvent/index.scss
@@ -55,7 +55,7 @@
       }
     }
   }
-  &__accomodation{
+  &__category{
     > h2 {
       margin: 0;
       border-bottom: $dash-border;
