+++ shared/components/SpeakerList/index.scss
