+++ shared/components/SpeakerList/test.js
